requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.6.2
apache-beam==2.46.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autopep8==2.0.1
avro-python3==1.10.2
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
boto3==1.26.106
botocore==1.29.106
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
crcmod==1.7
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.6
dm-tree==0.1.8
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
etils==1.1.1
executing==1.2.0
fast_ctc_decode==0.3.0
fastapi==0.92.0
fastavro==1.7.3
fasteners==0.18
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.39.2
fqdn==1.5.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
hdfs==2.7.0
httplib2==0.21.0
httptools==0.5.0
huggingface-hub==0.5.1
idna==3.4
imagecodecs==2022.2.22
imageio==2.15.0
immutabledict==2.2.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jstyleson==0.0.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
kaggle==1.5.13
keras==2.11.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lmdb==1.3.0
lvis==0.5.3
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
motor==3.1.1
munkres==1.1.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.7.1
nibabel==3.2.2
nltk==3.7
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///Users/andradel/Documents/studies/python/machine_learning/custom_object_detection/models/research
objsize==0.6.1
opencv-contrib-python==4.5.5.64
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
openvino==2022.1.0
openvino-dev==2022.1.0
openvino-telemetry==2022.1.1
opt-einsum==3.3.0
orjson==3.8.8
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parasail==1.2.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
portalocker==2.7.0
progress==1.6
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyclipper==1.3.0.post2
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pydicom==2.3.0
pydot==1.4.2
Pygments==2.14.0
pymongo==3.13.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
rawpy==0.17.1
regex==2023.3.22
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.2.0
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.97
seqeval==1.2.2
Shapely==1.8.1.post1
six==1.15.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.25.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-macos==2.11.0
tensorflow-metadata==1.12.0
tensorflow-metal==0.7.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.11.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.11.4
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.3.25
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.16.2
typeguard==3.0.2
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
Wave==0.0.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.6
wrapt==1.14.1
zipp==3.11.0
zstandard==0.20.0
